pyproject.toml
[project]
name = "aios"
version = "0.2.0"
description = "AIOS — Native AI OS, AIOS-specific layer on top of nanobot-ai"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "aoye516" }]

# Hard runtime deps for AIOS-specific functionality.
# nanobot itself is installed separately as editable: `uv pip install -e vendor/nanobot`
dependencies = [
    "asyncpg>=0.29",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
aios = "aios.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aios"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
